# manager/cloudhv/openapi.yaml
openapi: 3.0.1
info:
  title: Cloud Hypervisor API
  description: Local HTTP based API for managing and inspecting a cloud-hypervisor virtual machine.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.3.0
servers:
  - url: http://localhost/api/v1
paths:
...
  /vm.create:
    put:
      summary: Create the cloud-hypervisor Virtual Machine (VM) instance. The instance is not booted, only created.
      operationId: createVM
      requestBody:
        description: The VM configuration
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VmConfig"
        required: true
      responses:
        204:
          description: The VM instance was successfully created.
...
components:
  schemas:
...
    PayloadConfig:
      type: object
      properties:
        firmware:
          type: string
        kernel:
          type: string
        cmdline:
          type: string
        initramfs:
          type: string
        igvm:
          type: string
        host_data:
          type: string
      description: Payloads to boot in guest
    VmConfig:
      required:
        - payload
      type: object
      properties:
        cpus:
          $ref: "#/components/schemas/CpusConfig"
        memory:
          $ref: "#/components/schemas/MemoryConfig"
        payload:
          $ref: "#/components/schemas/PayloadConfig"
        rate_limit_groups:
          type: array
          items:
            $ref: "#/components/schemas/RateLimitGroupConfig"
        disks:
          type: array
          items:
            $ref: "#/components/schemas/DiskConfig"
        net:
          type: array
          items:
            $ref: "#/components/schemas/NetConfig"
        rng:
          $ref: "#/components/schemas/RngConfig"
        balloon:
          $ref: "#/components/schemas/BalloonConfig"
        fs:
          type: array
          items:
            $ref: "#/components/schemas/FsConfig"
        generic-vhost-user:
          type: array
          items:
            $ref: "#/components/schemas/GenericVhostUserConfig"
        pmem:
          type: array
          items:
            $ref: "#/components/schemas/PmemConfig"
        serial:
          $ref: "#/components/schemas/ConsoleConfig"
        console:
          $ref: "#/components/schemas/ConsoleConfig"
        debug_console:
          $ref: "#/components/schemas/DebugConsoleConfig"
        devices:
          type: array
          items:
            $ref: "#/components/schemas/DeviceConfig"
        vdpa:
          type: array
          items:
            $ref: "#/components/schemas/VdpaConfig"
        vsock:
          $ref: "#/components/schemas/VsockConfig"
        numa:
          type: array
          items:
            $ref: "#/components/schemas/NumaConfig"
        iommu:
          type: boolean
          default: false
        watchdog:
          type: boolean
          default: false
        pvpanic:
          type: boolean
          default: false
        pci_segments:
          type: array
          items:
            $ref: "#/components/schemas/PciSegmentConfig"
        platform:
          $ref: "#/components/schemas/PlatformConfig"
        tpm:
          $ref: "#/components/schemas/TpmConfig"
        landlock_enable:
          type: boolean
          default: false
        landlock_rules:
          type: array
          items:
            $ref: "#/components/schemas/LandlockConfig"
      description: Virtual machine configuration
    CpuAffinity:
      required:
        - vcpu
        - host_cpus
      type: object
      properties:
        vcpu:
          type: integer
        host_cpus:
          type: array
          items:
            type: integer
    CpuFeatures:
      type: object
      properties:
        amx:
          type: boolean
    CpuTopology:
      type: object
      properties:
        threads_per_core:
          type: integer
        cores_per_die:
          type: integer
        dies_per_package:
          type: integer
        packages:
          type: integer
    CoreSchedulingMode:
      type: string
      enum: ["Vm", "Vcpu", "Off"]
      default: "Vm"
    CpusConfig:
      required:
        - boot_vcpus
        - max_vcpus
      type: object
      properties:
        boot_vcpus:
          minimum: 1
          type: integer
        max_vcpus:
          minimum: 1
          type: integer
        topology:
          $ref: "#/components/schemas/CpuTopology"
        kvm_hyperv:
          type: boolean
          default: false
        max_phys_bits:
          type: integer
        nested:
          type: boolean
          default: true
        affinity:
          type: array
          items:
            $ref: "#/components/schemas/CpuAffinity"
        features:
          $ref: "#/components/schemas/CpuFeatures"
        core_scheduling:
          $ref: "#/components/schemas/CoreSchedulingMode"
    PciSegmentConfig:
      required:
        - pci_segment
      type: object
      properties:
        pci_segment:
          type: integer
          format: int16
        mmio32_aperture_weight:
          type: integer
          format: int32
        mmio64_aperture_weight:
          type: integer
          format: int32
    PlatformConfig:
      type: object
      properties:
        num_pci_segments:
          type: integer
          format: int16
        iommu_segments:
          type: array
          items:
            type: integer
            format: int16
        iommu_address_width:
          type: integer
          format: uint8
        serial_number:
          type: string
        uuid:
          type: string
        oem_strings:
          type: array
          items:
            type: string
        tdx:
          type: boolean
          default: false
        sev_snp:
          type: boolean
          default: false
    MemoryZoneConfig:
      required:
        - id
        - size
      type: object
      properties:
        id:
          type: string
        size:
          type: integer
          format: int64
        file:
          type: string
        mergeable:
          type: boolean
          default: false
        shared:
          type: boolean
          default: false
        hugepages:
          type: boolean
          default: false
        hugepage_size:
          type: integer
          format: int64
        host_numa_node:
          type: integer
          format: int32
        hotplug_size:
          type: integer
          format: int64
        hotplugged_size:
          type: integer
          format: int64
        prefault:
          type: boolean
          default: false
    MemoryConfig:
      required:
        - size
      type: object
      properties:
        size:
          type: integer
          format: int64
        hotplug_size:
          type: integer
          format: int64
        hotplugged_size:
          type: integer
          format: int64
        mergeable:
          type: boolean
          default: false
        hotplug_method:
          type: string
          default: "Acpi"
        shared:
          type: boolean
          default: false
        hugepages:
          type: boolean
          default: false
        hugepage_size:
          type: integer
          format: int64
        prefault:
          type: boolean
          default: false
        thp:
          type: boolean
          default: true
        zones:
          type: array
          items:
            $ref: "#/components/schemas/MemoryZoneConfig"
...
    RateLimitGroupConfig:
      required:
        - id
        - rate_limiter_config
      type: object
      properties:
        id:
          type: string
        rate_limiter_config:
          $ref: "#/components/schemas/RateLimiterConfig"
...
    ImageType:
      type: string
      enum: ["FixedVhd", "Qcow2", "Raw", "Vhdx", "Unknown"]
...
    DiskConfig:
      type: object
      properties:
        path:
          type: string
        readonly:
          type: boolean
          default: false
        direct:
          type: boolean
          default: false
        iommu:
          type: boolean
          default: false
        num_queues:
          type: integer
          default: 1
        queue_size:
          type: integer
          default: 128
        vhost_user:
          type: boolean
          default: false
        vhost_socket:
          type: string
        rate_limiter_config:
          $ref: "#/components/schemas/RateLimiterConfig"
        pci_segment:
          type: integer
          format: int16
        id:
          type: string
        serial:
          type: string
        rate_limit_group:
          type: string
        queue_affinity:
          type: array
          items:
            $ref: "#/components/schemas/VirtQueueAffinity"
        backing_files:
          type: boolean
          default: false
        sparse:
          type: boolean
          default: true
        image_type:
          $ref: "#/components/schemas/ImageType"
        lock_granularity:
          $ref: "#/components/schemas/LockGranularity"
    NetConfig:
      type: object
      properties:
        tap:
          type: string
        ip:
          type: string
          default: "192.168.249.1"
          description: IPv4 or IPv6 address
        mask:
          type: string
          default: "255.255.255.0"
          description: Must be a valid IPv4 netmask if ip is an IPv4 address or a valid IPv6 netmask if ip is an IPv6 address.
        mac:
          type: string
        host_mac:
          type: string
        mtu:
          type: integer
        iommu:
          type: boolean
          default: false
        num_queues:
          type: integer
          default: 2
        queue_size:
          type: integer
          default: 256
        vhost_user:
          type: boolean
          default: false
        vhost_socket:
          type: string
        vhost_mode:
          type: string
          default: "Client"
        id:
          type: string
        pci_segment:
          type: integer
          format: int16
        rate_limiter_config:
          $ref: "#/components/schemas/RateLimiterConfig"
        offload_tso:
          type: boolean
          default: true
        offload_ufo:
          type: boolean
          default: true
        offload_csum:
          type: boolean
          default: true
    RngConfig:
      required:
        - src
      type: object
      properties:
        src:
          type: string
        iommu:
          type: boolean
          default: false
    BalloonConfig:
      required:
        - size
      type: object
      properties:
        size:
          type: integer
          format: int64
        deflate_on_oom:
          type: boolean
          default: false
          description: Deflate balloon when the guest is under memory pressure.
        free_page_reporting:
          type: boolean
          default: false
          description: Enable guest to report free pages.
    FsConfig:
      required:
        - num_queues
        - queue_size
        - socket
        - tag
      type: object
      properties:
        tag:
          type: string
        socket:
          type: string
        num_queues:
          type: integer
          default: 1
        queue_size:
          type: integer
          default: 1024
        pci_segment:
          type: integer
          format: int16
        id:
          type: string
    GenericVhostUserConfig:
      required:
        - queue_sizes
        - socket
        - tag
        - virtio_id
      type: object
      properties:
        socket:
          type: string
        queue_size:
          type: array
          items:
            type: uint16
        pci_segment:
          type: integer
          format: int16
        virtio_id:
          type: uint32
    PmemConfig:
      required:
        - file
      type: object
      properties:
        file:
          type: string
        size:
          type: integer
          format: int64
        iommu:
          type: boolean
          default: false
        discard_writes:
          type: boolean
          default: false
        pci_segment:
          type: integer
          format: int16
        id:
          type: string
...
    ConsoleConfig:
      required:
        - mode
      type: object
      properties:
        file:
          type: string
        socket:
          type: string
        mode:
          $ref: "#/components/schemas/ConsoleMode"
        iommu:
          type: boolean
          default: false
    DebugConsoleConfig:
      required:
        - mode
      type: object
      properties:
        file:
          type: string
        mode:
          $ref: "#/components/schemas/ConsoleMode"
        iobase:
          type: integer
    DeviceConfig:
      required:
        - path
      type: object
      properties:
        path:
          type: string
        iommu:
          type: boolean
          default: false
        pci_segment:
          type: integer
          format: int16
        id:
          type: string
        x_nv_gpudirect_clique:
          type: integer
          format: int8
    TpmConfig:
      required:
        - socket
      type: object
      properties:
        socket:
          type: string
    VdpaConfig:
      required:
        - path
        - num_queues
      type: object
      properties:
        path:
          type: string
        num_queues:
          type: integer
          default: 1
        iommu:
          type: boolean
          default: false
        pci_segment:
          type: integer
          format: int16
        id:
          type: string
    VsockConfig:
      required:
        - cid
        - socket
      type: object
      properties:
        cid:
          type: integer
          format: int64
          minimum: 3
          description: Guest Vsock CID
        socket:
          type: string
          description: Path to UNIX domain socket, used to proxy vsock connections.
        iommu:
          type: boolean
          default: false
        pci_segment:
          type: integer
          format: int16
        id:
          type: string
...
    NumaConfig:
      required:
        - guest_numa_id
      type: object
      properties:
        guest_numa_id:
          type: integer
          format: int32
        cpus:
          type: array
          items:
            type: integer
            format: int32
        distances:
          type: array
          items:
            $ref: "#/components/schemas/NumaDistance"
        memory_zones:
          type: array
          items:
            type: string
        pci_segments:
          type: array
          items:
            type: integer
            format: int32
        device_id:
          type: string
...
    LandlockConfig:
      required:
        - path
        - access
      type: object
      properties:
        path:
          type: string
        access:
          type: string